# docs/api/openapi.yaml
openapi: 3.1.1
info:
  title: Sick-Fansubs HTTP API
  version: "1"
  description: >-
    Accepted Sick-Fansubs wire contracts. Product APIs use /api/v1; operational
    health endpoints are intentionally outside product API versioning.

    Cross-cutting: every endpoint whose chain carries the forced-change gate
    (the authenticated reads chain and the authenticated chain) additionally
    answers 403 with problem type /problems/auth/password-change-required for
    sessions whose account carries the forced-change flag. The /api/v1/auth/
    subtree never carries the gate, so its paths stay reachable while flagged —
    including PUT /api/v1/auth/password, the sign-out pair, the session
    bootstrap, POST /api/v1/auth/reset-password, and
    POST /api/v1/auth/resend-verification.

    Cross-cutting: every documented unsafe operation (POST, PUT, PATCH,
    DELETE) requires the trusted origin and, when a session is present, the
    session-bound X-CSRF-Token header matching the csrfToken returned by the
    session bootstrap — a missing, malformed, or mismatched token is a 403.
    Safe methods never require it, and an unauthenticated unsafe request
    (for example sign-in, register, forgot-password, or reset-password) is
    protected by the trusted origin alone. An expired session has the
    clearing sf_session cookie written before its handler runs, and the
    operations whose handler clears the cookie — sign-out, sign-out-all,
...
    their own response.

...
servers:
...
paths:
...
  /api/v1/users/{id}/suspend:
    post:
      operationId: suspendUser
      summary: Suspend a user account (staff)
      description: >-
        Peer-protected suspension: sets status = suspended, bumps
        auth_version, and deletes the target's sessions in one transaction,
        so a suspended account cannot sign in and loses every live session.
        No content, role, or profile changes occur. Floor: moderator+ with
        peer protection — super-admins may suspend anyone including
        themselves; admins may suspend moderators and users; moderators may
        suspend users only. Suspending the last ACTIVE super-admin —
        including self-suspension — is a 409 with problem type
        /problems/auth/last-super-admin. A same-status POST is a 200 no-op
        (no revocation, not audited) and sits behind authorization. Blocked
        403s and guarded 409s are audited as user_suspended failures; the
        success audit carries the target's role snapshot, and a failed write
        (500) is audited as a user_suspended failure.
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: Account suspended (or the no-op) — the updated projection.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/LastSuperAdminConflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /api/v1/users/{id}/reactivate:
    post:
      operationId: reactivateUser
      summary: Reactivate a user account (staff)
      description: >-
        Peer-protected reactivation: sets status = active and bumps
        auth_version, requiring a new sign-in (suspension already deleted the
        sessions). Floor and peer protection mirror the suspend endpoint.
        Reactivation can never reduce the active super-admin count, so the
        last-active-super-admin guard does not apply. A same-status POST is a
        200 no-op (no revocation, not audited) and sits behind authorization.
        Blocked 403s are audited as user_reactivated failures; the success
        audit carries the target's role snapshot, and a failed write (500) is
        audited as a user_reactivated failure.
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "200":
          description: Account reactivated (or the no-op) — the updated projection.
          headers:
            X-Request-ID:
              $ref: "#/components/headers/RequestID"
            Cache-Control:
              $ref: "#/components/headers/NoStoreCacheControl"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffUser"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/InvalidCredentials"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  # ── User deletion ──
...
components:
  headers:
    RequestID:
      description: Server-generated request correlation identifier.
      schema:
        type: string
    NoStoreCacheControl:
      description: API responses default to no-store.
      schema:
        type: string
        const: no-store
  parameters:
...
    UserID:
      name: id
      in: path
      required: true
      description: >-
        Opaque user identifier (1-64 RFC 3986 unreserved characters). An id
        outside this shape is a 400 — structurally invalid input, distinct
        from the masked 404.
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[A-Za-z0-9._~-]{1,64}$"
...
  schemas:
...
    StaffUser:
      type: object
      description: >-
        One account in the staff user list.
        email is present only when the viewer holds admin or above — a
        moderator's response omits the key entirely. avatarUrl is NOT
        role-gated: it is always present, null when the account has no
        avatar. createdAt is a canonical API instant.
      additionalProperties: false
      required:
        - id
        - username
        - role
        - status
        - emailVerified
        - createdAt
        - avatarUrl
      properties:
        id:
          type: string
          description: Opaque user identifier.
        username:
          type: string
        email:
          type: string
          description: >-
            Lowercased account email. Present only for
            admin+ viewers.
        role:
          type: string
          description: Accepted role identifier.
        status:
          type: string
          enum:
            - active
            - suspended
        emailVerified:
          type: boolean
          description: Whether the account proved inbox control.
        createdAt:
          type: string
          format: date-time
          description: Account creation instant, canonical UTC RFC 3339 with milliseconds.
        avatarUrl:
          type:
            - string
            - "null"
          description: >-
            ABSOLUTE URL of the served avatar image (the 2-hex storage
            subdirectory never appears on the wire), served through
            GET /media/images/{file} with immutable caching, or null when
            the account has no avatar.
...
    Problem:
      type: object
      description: >-
        RFC 9457 Problem Details with only accepted safe extensions. Endpoint
        contracts define their stable problem types and violation codes.
      additionalProperties: false
      required:
        - type
        - title
        - status
        - requestId
      properties:
        type:
          type: string
          format: uri-reference
          description: Stable relative problem type identifier.
          example: /problems/validation
        title:
          type: string
          description: Safe developer-oriented summary; not localized UI copy.
        status:
          type: integer
          minimum: 100
          maximum: 599
          description: HTTP status code for this occurrence.
        requestId:
          type: string
          description: Server-generated request correlation identifier.
        violations:
          type: array
          description: Deterministically ordered field violations when applicable.
          items:
            type: object
            additionalProperties: false
            required:
              - field
              - code
            properties:
              field:
                type: string
                description: Accepted public request-field identifier.
              code:
                type: string
                description: Stable machine-readable violation identifier.
...
  responses:
    BadRequest:
      description: Malformed request (400) — problem type /problems/bad-request.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    InvalidCredentials:
      description: Generic 401 — problem type /problems/auth/invalid-credentials.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    Forbidden:
      description: >-
        403 — problem type /problems/forbidden (CSRF/origin failure or a role
        floor) or /problems/auth/password-change-required (the forced-change
        gate — see PasswordChangeRequired).
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    LastSuperAdminConflict:
      description: >-
        409 — problem type /problems/auth/last-super-admin: the operation
        (role demotion, deletion, or suspension) would leave zero ACTIVE
        super-admin accounts.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    NotFound:
      description: >-
        404 — problem type /problems/not-found: an unknown path under /api/v1
        (namespace catch-all), or an absent target on a resource operation
        (the operation's description states what its 404 discloses).
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    InternalError:
      description: 500 generic unexpected failure — problem type /problems/internal-error.
      headers:
        X-Request-ID:
          $ref: "#/components/headers/RequestID"
        Cache-Control:
          $ref: "#/components/headers/NoStoreCacheControl"
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"